fixing issues with user subscription graph
creating subscription graph using client side code
