+++ MrCMS.Web/Areas/Admin/Services/UserSubscriptionReports/IUserSubscriptionReportsService.cs
@@ -6,8 +6,8 @@ using MrCMS.Web.Areas.Admin.Models.UserSubscriptionReports;
 
 namespace MrCMS.Web.Areas.Admin.Services.UserSubscriptionReports
 {
-    interface IUserSubscriptionReportsService
+    public interface IUserSubscriptionReportsService
     {
-        IEnumerable<JsonObject> GetAllSubscriptions(UserSubscriptionReportsSearchQuery searchQuery);
+        IEnumerable<object> GetAllSubscriptions(UserSubscriptionReportsSearchQuery searchQuery);
     }
 }
